pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-vk-repost-bot"
version = "0.1.0"
description = "Telegram to VK repost bot"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "SQLAlchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.1.0",
  "celery>=5.3.0",
  "redis>=5.0.0",
]

[tool.setuptools]
packages = ["app"]
